refactor: update home page component file naming
