+++ src/components/(marketing)/home/main-discord-embed.tsx
